Make SheepDyeWoolEvent extend generic EntityDyeEvent.
